Ensure configuration module is compiled
We need the configuration module to be compiled when we expand the
intercept macros, and this is the only way we can do it. This is a sync
operation that will ask the code server for the configuration module to
be compiled, and will block the compilation of the current module while
that doesn't happen.

Since this interaction only happens in compile-time, it doesn't matter.
If the `Code.ensure_compiled/1` was being called in runtime, we would
have a problem.