Configuration needs MFA of the callback functions
Instead of the `{module, function}` that we needed on the previous
version. We know by definition the arity of the callback functions, but
this change both simplifies the configuration DSL and unifies how we
point to functions: both intercepted functions and callback functions
are now referred to as `{module, function, args}`.

Updated the documentation as well.