Override global configuration with `use Interceptor, config: <CONFIG>`
`<CONFIG>` may be a module that exposes the `get_intercept_config/0`
function, or it may be an intercept configuration map, using MFA strings
(i.e. `"Module.function/arity"`) or MFA in tuple format (i.e. `{module,
function, arity}`).

Inside the `Interceptor`, now we have the concept of a "global"
configuration, coming from the `Application.get_env/2`, and the "own"
config, that may exist or not, and it's merged on top of the former.